requirements.txt
azure-functions==1.21.3
pyodbc==5.1.2
azure-identity==1.17.0
